pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "verizon_connect_api"
description = "Helper class for Verizon Connect API"
readme = "README.md"
license = { text = "MIT" }
version = "0.3.0"
requires-python = ">=3.8"
authors = [
    {name = "Ethan O'Leske", email = "ethan.oleske@gmail.com"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "requests~=2.32"
]

[project.optional-dependencies]
test = [
    "python-dotenv~=1.0",
    "pydantic~=2.8"
]
doc = [
    "sphinx",
    "sphinx_rtd_theme",
    "sphinx_pyproject",
    "m2r"
]

[project.urls]
Homepage = "https://github.com/edoleske/py-verizon-connect-api"
Documentation = "https://edoleske.github.io/py-verizon-connect-api"

[tool.pytest.ini_options]
pythonpath = "src"
addopts = [
    "--import-mode=importlib"
]
